improved welding in editmode, added more error handlers

+++ 2.7/Welder.py
@@ -29,7 +29,7 @@ from bpy_extras.view3d_utils import (
 bl_info = {
     "name": "Welder",
     "author": "Łukasz Hoffmann",
-    "version": (0,0, 7),
+    "version": (0,0, 8),
     "location": "View 3D > Object Mode > Tool Shelf",
     "blender": (2, 7, 9),
     "description": "Generate weld along the odge of intersection of two objects",
@@ -48,8 +48,11 @@ class WelderDrawOperator(bpy.types.Operator):
     bl_label = "Draw"   
     
     def modal(self, context, event):
-        if (context.area==None or bpy.context.object==None or context.space_data.type!='VIEW_3D'):
-            self.unregister_handlers(context)
+        if (context.area==None or bpy.context.object==None):
+            try:
+                self.unregister_handlers(context)
+            except:
+                "Finished"    
             bpy.context.scene.welddrawing=False
             for km in self.list: km.active = True
             return {'CANCELLED'}    
@@ -255,27 +258,40 @@ class OBJECT_OT_WeldButton(bpy.types.Operator):
     
     
     def execute(self, context):
+        preserve=True
+        objectstodel=bpy.context.selected_objects
         if (bpy.context.object==None):
             self.report({'ERROR'}, 'Invalid context or nothing selected')
             return {'FINISHED'}    
         if (bpy.context.object.mode=='EDIT'):
             if (bpy.context.scene.objects.active.type=='CURVE'):
                 bpy.ops.object.mode_set(mode = 'OBJECT')
-            else:    
-                bpy.ops.mesh.duplicate()
-                bpy.ops.mesh.separate(type='SELECTED')
-                bpy.ops.object.mode_set(mode='OBJECT')
-                originobj=bpy.context.scene.objects.active
-                obj=bpy.context.selected_objects[0]
-                bpy.context.scene.objects.active = obj
-                originobj.select=False  
+            else:                    
+                if bpy.context.scene.objects.active.type=='MESH':
+                    if absoluteselection(bpy.context.scene.objects.active):                         
+                        preserve=False
+                        objectstodel=bpy.context.selected_objects
+                        for o in objectstodel: print(o.name)
+                    else:
+                        if (not isanythingselected(bpy.context.scene.objects.active)):
+                            self.report({'ERROR'}, 'Nothing selected, aborting')
+                            return {'FINISHED'}     
+                        bpy.ops.mesh.duplicate()
+                        bpy.ops.mesh.separate(type='SELECTED')
+                        bpy.ops.object.mode_set(mode='OBJECT')
+                        originobj=bpy.context.scene.objects.active
+                        obj=bpy.context.selected_objects[0]
+                        bpy.context.scene.objects.active = obj
+                        originobj.select=False  
+                        
+                        if (obj.type=='MESH' and (len(obj.data.polygons)>0 or not iscontinuable(obj))):
+                            bpy.ops.object.delete()
+                            bpy.context.scene.objects.active=originobj
+                            bpy.ops.object.mode_set(mode='EDIT')
+                            self.report({'ERROR'}, 'Detected wrong selectiong or not an edgeloop, aborting')
+                            return {'FINISHED'}    
                 
-                if (obj.type=='MESH' and (len(obj.data.polygons)>0 or not iscontinuable(obj))):
-                    bpy.ops.object.delete()
-                    bpy.context.scene.objects.active=originobj
-                    bpy.ops.object.mode_set(mode='EDIT')
-                    self.report({'ERROR'}, 'Detected selected faces or not an edgeloop, aborting')
-                    return {'FINISHED'}                  
+                              
         if (bpy.context.object.mode!='OBJECT'):
             self.report({'ERROR'}, 'Welding works only in edit or object mode')
             return {'FINISHED'}
@@ -397,8 +413,7 @@ class OBJECT_OT_WeldButton(bpy.types.Operator):
             for f in bm1.edges:
                 f.select=not f.select
 
-            bpy.ops.mesh.delete(type='EDGE') # usuwanie wielokątow    
-
+            bpy.ops.mesh.delete(type='EDGE') # usuwanie wielokątow
 
             #POPRAWKA DO SKRYPTU MANIFOLD - USUWANIE TROJKATOW PRZY KRAWEDZIACH
             bpy.ops.mesh.select_mode(type="EDGE")
@@ -435,7 +450,105 @@ class OBJECT_OT_WeldButton(bpy.types.Operator):
             
             MakeWeldFromCurve(OBJ1,edge_length,obje,matrix)
             
+            if not preserve:
+                removebyobj(OBJ3)
+                removebyobj(OBJ4)
+            
             return bpy.ops.weld.translate('INVOKE_DEFAULT')
+
+def isanythingselected(obj):
+    bm=bmesh.from_edit_mesh(obj.data)
+    vertices=[v.index for v in bm.verts if v.select]
+    faces=[f.index for f in bm.faces if f.select]
+    edges=[e.index for e in bm.edges if e.select]
+    selection=len(vertices)+len(faces)+len(edges)
+    if selection==0: return False
+    else: return True
+
+def removebyobj(obj):
+    oldselection=bpy.context.selected_objects
+    oldactive=bpy.context.scene.objects.active
+    bpy.ops.object.select_all(action = 'DESELECT')
+    obj.select=True
+    bpy.context.scene.objects.active=obj
+    bpy.ops.object.delete()
+    bpy.context.scene.objects.active=oldactive
+    for o in oldselection:
+        o.select=True
+    
+def is_point_inside(point,ob):
+    axes = [ mathutils.Vector((1,0,0)) ]
+    outside = False
+    for axis in axes:
+        mat = ob.matrix_world
+        mat.invert()
+        #mat = mathutils.Matrix(ob.matrix_world).invert()
+        orig = mat*point
+        mat.invert()
+        count = 0
+        while True:
+            hit,location,normal,index = ob.ray_cast(orig,orig+axis*10000.0)
+            if index == -1: break
+            count += 1
+            orig = location + axis*0.00001
+        if count%2 == 0:
+            outside = True
+            break
+    return not outside
+
+def mesh_intersecting(obj1,obj2):
+    intersection=False
+    for v in obj1.data.vertices:
+        if is_point_inside(v.co,obj2):
+            intersection=True
+            break;
+    return intersection
+
+def absoluteselection(obj):
+    bm=bmesh.from_edit_mesh(obj.data)
+    selected_old_indices=[v.index for v in bm.verts if v.select]
+    selected_old_faces=[f.index for f in bm.faces if f.select]
+    selected_old_edges=[e.index for e in bm.edges if e.select]
+    selected_old=len(selected_old_indices)
+    if selected_old==0: return False
+    bpy.ops.mesh.select_more()
+    selected_new=len([v.index for v in bm.verts if v.select])
+    if selected_new==selected_old:
+        bpy.ops.mesh.duplicate()
+        bpy.ops.mesh.separate(type='SELECTED')        
+        bpy.ops.object.mode_set(mode='OBJECT')        
+        obj_new=bpy.context.selected_objects[0]
+        for o in bpy.context.selected_objects:
+            if o!=bpy.context.scene.objects.active:
+                obj_new=o
+        bpy.ops.object.select_all(action = 'DESELECT')            
+        bpy.context.scene.objects.active=obj_new
+        obj_new.select=True
+        bpy.ops.object.mode_set(mode='EDIT')
+        bpy.ops.mesh.separate(type='LOOSE')
+        bpy.ops.object.mode_set(mode='OBJECT')
+        bpy.context.scene.objects.active.select
+        bpy.ops.object.mode_set(mode='OBJECT')
+        if len(bpy.context.selected_objects)>1:
+            bpy.ops.object.transform_apply(location=True, rotation=True, scale=True)
+            if mesh_intersecting(bpy.context.selected_objects[0],bpy.context.selected_objects[1]):
+                return True, None, None
+        bpy.context.scene.objects.active.select
+        bpy.ops.object.delete()
+        obj_new.select
+        bpy.context.scene.objects.active=obj_new
+        bpy.ops.object.delete()
+    for f in bm.faces:
+        if f.index in selected_old_faces: f.select=True
+        else: f.select=False
+    for f in bm.edges:
+        if f.index in selected_old_edges: f.select=True
+        else: f.select=False
+    for v in bm.verts:
+        if v.index in selected_old_indices: 
+            v.select=True
+        else: v.select=False           
+    return False    
 
 def iscontinuable(obj):
     continuable=True
